videoplayer: fix an integer overflow when seeking.

+++ src/api/qarvrecordedvideo.cpp
@@ -159,7 +159,8 @@ QArvDecoder* QArvRecordedVideo::makeDecoder() {
     }
 }
 
-bool QArvRecordedVideo::seek(uint frame) {
+bool QArvRecordedVideo::seek(quint64 frame)
+{
     return videofile.seek(frame*frameBytes_);
 }
